refactor: 할인금액 계산 클래스 분리
- 테스트 용이를 위한 클래스 분리
- 로직 변경에 의한 코드 수정
