+++ src/main/java/christmas/controller/EventPlannerController.java
@@ -2,6 +2,7 @@ package christmas.controller;
 
 import christmas.handler.ExceptionRetryHandler;
 import christmas.model.Benefits;
+import christmas.model.DiscountCalculator;
 import christmas.model.EventTotalBenefit;
 import christmas.model.EventValidator;
 import christmas.model.OrderSheets;
@@ -30,7 +31,7 @@ public class EventPlannerController {
         outputView.printOrderedMenus(orderedMenus);
         outputView.printTotalPriceBeforeDiscount(orderedMenus);
         TotalPrice totalPrice = new TotalPrice(orderedMenus.calculateTotalPrice());
-        EventValidator eventValidator = new EventValidator(orderedMenus, visitDate, totalPrice);
+        EventValidator eventValidator = new EventValidator(new DiscountCalculator(orderedMenus, visitDate), totalPrice);
         outputView.printGiveAwayMenu(eventValidator.getGiveAway());
         Benefits benefits = eventValidator.getBenefits();
         outputView.printBenefits(benefits.getBenefitInfos());
